[TR-208] Move routing variables from select-issue-type to route-data
