+++ public/canjs/controls/select-issue-type/select-issue-type.js
@@ -2,9 +2,8 @@ import { StacheElement, type, ObservableObject, ObservableArray, value, queues }
 
 import {updateUrlParam, pushStateObservable} from "../../routing/state-storage.js";
 import { bitoviTrainingIssueData } from "../../../examples/bitovi-training.js";
-import { getSimplifiedIssueHierarchy } from "../../../stateful-data/jira-data-requests.js";
-import { mostCommonElement } from "../../../utils/array/array-helpers.js";
 
+import routeData from "../../routing/route-data.js";
 import "../status-filter.js";
 
 import SimpleTooltip from "../../ui/simple-tooltip/simple-tooltip";
@@ -34,13 +33,13 @@ document.body.append(RELEASES_TOOLTIP);
 class TypeSelectionDropdown extends StacheElement {
     static view = `
         {{# for(issueType of this.issueHierarchy) }}
-        <label class="px-4 py-2 block {{#eq(this.primaryIssueType, issueType.name)}}bg-blue-101{{else}}${hoverEffect}{{/eq}}"><input 
+        <label class="px-4 py-2 block {{#eq(this.routeData.primaryIssueType, issueType.name)}}bg-blue-101{{else}}${hoverEffect}{{/eq}}"><input 
             type="radio" 
             name="primaryIssueType" 
-            checked:from="eq(this.primaryIssueType, issueType.name)"
+            checked:from="eq(this.routeData.primaryIssueType, issueType.name)"
             on:change="this.onSelection(issueType.name)"/> {{issueType.name}}s </label>
         {{/ }}
-        <label class="px-4 py-2  block {{#eq(this.primaryIssueType, 'Release')}}bg-blue-101{{else}}${hoverEffect}{{/eq}} border-t border-t-2 border-t-neutral-301"
+        <label class="px-4 py-2  block {{#eq(this.routeData.primaryIssueType, 'Release')}}bg-blue-101{{else}}${hoverEffect}{{/eq}} border-t border-t-2 border-t-neutral-301"
             on:mouseenter="this.showReleases(scope.element)">
             Releases <img class="inline" src="/images/chevron-right-new.svg"/> 
         </label>
@@ -49,7 +48,7 @@ class TypeSelectionDropdown extends StacheElement {
         let dropdown = new ReleasesTypeSelectionDropdown().initialize({
             issueHierarchy: this.issueHierarchy,
             onSelection: this.onSelection,
-            secondaryIssueType: this.secondaryIssueType
+            secondaryIssueType: this.routeData.secondaryIssueType
         })
 
         RELEASES_TOOLTIP.rightOfElementInScrollingContainer(label, dropdown);
@@ -61,10 +60,10 @@ customElements.define("select-type-dropdown", TypeSelectionDropdown);
 class ReleasesTypeSelectionDropdown extends StacheElement {
     static view = `
         {{# for(issueType of this.issueHierarchy) }}
-        <label class="px-4 py-2 block {{#eq(this.secondaryIssueType, issueType.name)}}bg-blue-101{{else}}${hoverEffect}{{/eq}}"><input 
+        <label class="px-4 py-2 block {{#eq(this.routeData.secondaryIssueType, issueType.name)}}bg-blue-101{{else}}${hoverEffect}{{/eq}}"><input 
             type="radio" 
             name="primaryIssueType" 
-            checked:from="eq(this.secondaryIssueType, issueType.name)"
+            checked:from="eq(this.routeData.secondaryIssueType, issueType.name)"
             on:change="this.onSelection('Release', issueType.name)"/> {{issueType.name}}s </label>
         {{/ }}
     `
@@ -77,132 +76,43 @@ customElements.define("select-release-type-dropdown", ReleasesTypeSelectionDropd
 export class SelectIssueType extends StacheElement {
     static view = `
         <label for="reportOn" class="${DROPDOWN_LABEL}">Report on</label>
-        {{# not(this.primaryIssueType) }}
+        {{# not(this.routeData.primaryIssueType) }}
             <button class="rounded bg-neutral-201 px-3 py-1" id="reportOn">Loading ... </button>
         {{/ }}
-        {{# if(this.primaryIssueType) }}
+        {{# if(this.routeData.primaryIssueType) }}
             <button class="rounded bg-neutral-201 px-3 py-1 ${hoverEffect}" 
                 on:click="this.showChildOptions()" 
                 id="reportOn">
-                {{this.primaryIssueType}}s
-                {{# if(this.secondaryIssueType) }} / {{this.secondaryIssueType}}s {{/ if }}
+                {{this.routeData.primaryIssueType}}s
+                {{# if(this.routeData.secondaryIssueType) }} / {{this.routeData.secondaryIssueType}}s {{/ if }}
                 <img class="inline" src="/images/chevron-down.svg"/>
             </button>
         {{/ }}
     `;
-    static props ={
-        simplifiedIssueHierarchy: {
-            async(){
-                return getSimplifiedIssueHierarchy({
-                    isLoggedIn: this.jiraHelpers.hasValidAccessToken(),
-                    jiraHelpers: this.jiraHelpers,
-                });
-            }
-
+    static props ={     
+        routeData: {
+          get default() { return routeData; }
         },
-        
         get issueHierarchy(){
             return this.derivedIssues && this.derivedIssues.length ?
                 issueHierarchyFromNormalizedIssues(this.derivedIssues) :
                 this.simplifiedIssueHierarchy;
             
         },
-        // needs to be atomic
-        // a value like `Initiative`
-        // or `Release-Initiative`
-        selectedIssueType: {
-            value({resolve, lastSet, listenTo}) {
-                function getParamValue(){
-                    return new URL(window.location).searchParams.get("selectedIssueType") || "";
-                }
-                let timers = [];
-                function clearTimers() {
-                    timers.forEach((value) => clearTimeout(value));
-                    timers = [];
-                }
-
-                // anything happens in state, update the route 
-                // the route updates, update the state (or the route if it's wrong)
-                const resolveCurrentValue = () => {
-                    clearTimers();
-                    const curParamValue = getParamValue();
-
-                    // we wait to resolve to a defined value until we can check it's right
-                    if(this.issueHierarchy && this.issueHierarchy.length) {
-                        const curParamValue = getParamValue();
-
-                        // helps with legacy support to pick the first type
-                        if(curParamValue === "Release") {
-                            resolve( "Release-"+this.issueHierarchy[0].name );
-                        } else {
-                            const curSelectedParts = toSelectedParts(curParamValue);
-                            //const lastSelectedParts = toSelectedParts(lastSelectedValue);
-
-                            if(curSelectedParts) {
-                                // check it's ok
-                                let typeToCheck = curSelectedParts.secondary ?? curSelectedParts.primary;
-                                
-                                if(this.issueHierarchy.some( issue => issue.name === typeToCheck ) ) {
-                                    // make sure we actually need to update
-                                    resolve(curParamValue);
-                                } 
-                                // set back to default
-                                else {
-                                    timers.push( setTimeout( ()=> {
-                                        updateUrlParam("selectedIssueType", "", "");
-                                    },20) );
-                                }
-
-                            } else {
-                                // default to the first type
-                                resolve( this.issueHierarchy[0].name );
-                            }
-                        }
-                    } else {
-                        resolve(undefined)
-                    }
-                }
-
-
-                // when the route changes, check stuff ...
-                listenTo(pushStateObservable, ()=>{
-                    resolveCurrentValue();
-                })
-                
-                listenTo("issueHierarchy",({value})=> {
-                    resolveCurrentValue();
-                });
-
-                listenTo(lastSet, (value)=>{
-                    console.log("LAST SET sit", value)
-                    updateUrlParam("selectedIssueType", value, "");
-                });
-
-                resolveCurrentValue();
-
-            }
-        },
-        get primaryIssueType() {
-            return this.selectedIssueType && toSelectedParts(this.selectedIssueType).primary;
-        },
-        get secondaryIssueType() {
-            return this.selectedIssueType && toSelectedParts(this.selectedIssueType).secondary;
-        }
-
     };
     onSelection(primaryType, secondaryType){
         if(secondaryType) {
-            this.selectedIssueType = "Release-"+secondaryType;
+            this.routeData.selectedIssueType = "Release-"+secondaryType;
         } else {
-            this.selectedIssueType = primaryType;
+            this.routeData.selectedIssueType = primaryType;
         }
         TOOLTIP.leftElement();
         RELEASES_TOOLTIP.leftElement();
     }
     showChildOptions(){
         let dropdown = new TypeSelectionDropdown().initialize({
-            primaryIssueType: this.primaryIssueType,
-            secondaryIssueType: this.secondaryIssueType,
+            primaryIssueType: this.routeData.primaryIssueType,
+            secondaryIssueType: this.routeData.secondaryIssueType,
             issueHierarchy: this.issueHierarchy,
             onSelection: this.onSelection.bind(this)
         })
@@ -219,34 +129,5 @@ export class SelectIssueType extends StacheElement {
     }
 }
 
-function toSelectedParts(value){
-    if(value) {
-        if(value.startsWith("Release-")) {
-            return {primary: "Release", secondary: value.substring("Release-".length)}
-        } else {
-            return {primary: value}
-        }
-    } else {
-        return undefined;
-    }
-}
-
-/**
- * 
- * @param {Array<import("../../../jira/normalized/normalize.js").NormalizedIssue>} normalizedIssues 
- * @returns {Array<{type: string, hierarchyLevel: number}>}
- */
-function issueHierarchyFromNormalizedIssues(normalizedIssues){
-    const levelsToNames = []
-    for( let issue of normalizedIssues) {
-        if(!levelsToNames[issue.hierarchyLevel]) {
-            levelsToNames[issue.hierarchyLevel] = [];
-        }
-        levelsToNames[issue.hierarchyLevel].push(issue.type)
-    }
-    return levelsToNames.map( (names, i) => {
-        return {name: mostCommonElement(names), hierarchyLevel: i}
-    }).filter( i => i ).reverse()
-}
 
 customElements.define("select-issue-type", SelectIssueType);
